requirements.txt
flask
openai
reportlab
python-dotenv
werkzeug